# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dawn-go API By Golang Gin",
        "title": "Dawn-go API By Gin",
        "termsOfService": "https://github.com/zzlpeter/dawn-go",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/assassin/task/task": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Get special task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Update special task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "任务key",
                        "name": "task_key",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "执行方法",
                        "name": "execute_func",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "任务描述",
                        "name": "desc",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "调度类型",
                        "name": "trigger",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "调度频次",
                        "name": "spec",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "执行参数",
                        "name": "args",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "是否有效",
                        "name": "is_valid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "额外信息(json格式)",
                        "name": "extra",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "执行状态",
                        "name": "status",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Add special task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务key",
                        "name": "task_key",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "执行方法",
                        "name": "execute_func",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "任务描述",
                        "name": "desc",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "调度类型",
                        "name": "trigger",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "调度频次",
                        "name": "spec",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "执行参数",
                        "name": "args",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "是否有效",
                        "name": "is_valid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "执行状态",
                        "name": "status",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "额外信息",
                        "name": "extra",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerResponse"
                        }
                    }
                }
            }
        },
        "/assassin/task/tasks": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Get tasks list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.SwaggerResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}